+++ src/precice/precice.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
+#include <precice/Exceptions.hpp>
 #include <precice/Participant.hpp>
 #include <precice/Version.h>
-#include <precice/exceptions.hpp>
 #include <precice/span.hpp>
