+++ src/sources.cmake
@@ -246,6 +246,7 @@ target_sources(preciceCore
     src/partition/ReceivedPartition.cpp
     src/partition/ReceivedPartition.hpp
     src/partition/SharedPointer.hpp
+    src/precice/Exceptions.hpp
     src/precice/Participant.cpp
     src/precice/Participant.hpp
     src/precice/Tooling.cpp
@@ -256,7 +257,6 @@ target_sources(preciceCore
     src/precice/config/ParticipantConfiguration.cpp
     src/precice/config/ParticipantConfiguration.hpp
     src/precice/config/SharedPointer.hpp
-    src/precice/exceptions.hpp
     src/precice/impl/CommonErrorMessages.hpp
     src/precice/impl/DataContext.cpp
     src/precice/impl/DataContext.hpp
@@ -351,10 +351,10 @@ target_sources(preciceCore
 
 set_property(TARGET precice PROPERTY PUBLIC_HEADER
     ${PROJECT_BINARY_DIR}/src/precice/Version.h
+    src/precice/Exceptions.hpp
     src/precice/Participant.hpp
     src/precice/Tooling.hpp
     src/precice/Types.hpp
-    src/precice/exceptions.hpp
     src/precice/precice.hpp
     src/precice/span.hpp
     )
